0.3.0
The development of an improved model of interaction with the player has begun . To do this, new versions of the fragment, ViewModel, and date class have been added. A new adapter for the track list has been implemented and interaction with the selected track has been redesigned (now the player page does not open immediately after clicking on it in the list).